move StoreInner into its own submodule
